Create basic_directory_changer.py
Hi, im Hugo from the Netherlands, and i'm trying to get the hang of python. This is the first thing I post, and to be honest, I have no idea what I'm doing.
Was hoping someone is bored enough to look at this very simple file searcher I tried to build. 
Maybe threading the searching of the different drives could speed things up? Maybe 'layer' architecture is stupid to begin with.

Thanks for helping the newb. =]